Go back to storing timestamps for items in feed/ streams.
Undoes the optimization added by dc5296d975ecbca3. The timestamp that is stored
with the item body appears to be the most recent crawl timestamp (for which the
body changed), thus it's not accurate.

In addition to switching things back, changes how we store stream items from
N 2-item tuples to 2 N-item tuples (i.e. parallel tuples of timestamps and
item IDs). This reduces the in memory overhead, making this whole change still
net negative for memory consumption (-21MB in RSIZE for my account).